Support nested keys in TensorDictModule
